ASoC-ath79: Add option to enable external MCLK
This adds a new kernel option SND_ATH79_SOC_USE_EXTERNAL_MCLK, which
allows for the use of an external MCLK. Currently the divider values
are only hardcoded for a MCLK of 2457600Hz and 22579200Hz.

Also there is currently no logic to switch an external PLL between
these two MCLK frequencies.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>